clean code + update drag and drop behaviour
