+++ src/assets/templates/styles.css
@@ -1,0 +1,328 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    background-color: #f5f5f5;
+    padding: 20px;
+}
+
+.container {
+    max-width: 1200px;
+    margin: 0 auto;
+}
+
+.header {
+    text-align: center;
+    margin-bottom: 30px;
+    background: white;
+    padding: 20px;
+    border-radius: 10px;
+    box-shadow: 0 2px 10px rgba(0,0,0,0.1);
+}
+
+.title {
+    font-size: 2.5rem;
+    color: #2c3e50;
+    margin-bottom: 10px;
+}
+
+.instructions {
+    font-size: 1.1rem;
+    color: #7f8c8d;
+}
+
+.audio-elements-zone {
+    text-align: center;
+    margin: 30px 0;
+    background: white;
+    padding: 20px;
+    border-radius: 10px;
+    box-shadow: 0 2px 10px rgba(0,0,0,0.1);
+}
+
+.audio-elements-zone h3 {
+    color: #2c3e50;
+    margin-bottom: 15px;
+    font-size: 1.3rem;
+}
+
+.volume-controls {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 10px;
+    margin-bottom: 20px;
+    padding: 10px;
+    background: #f8f9fa;
+    border-radius: 8px;
+    border: 1px solid #e9ecef;
+}
+
+.volume-controls label {
+    font-weight: 500;
+    color: #495057;
+}
+
+.volume-slider {
+    width: 120px;
+    height: 6px;
+    background: #ddd;
+    outline: none;
+    border-radius: 3px;
+}
+
+.volume-slider::-webkit-slider-thumb {
+    appearance: none;
+    width: 18px;
+    height: 18px;
+    background: #2196f3;
+    cursor: pointer;
+    border-radius: 50%;
+}
+
+#volumeValue {
+    font-weight: 500;
+    color: #2196f3;
+    min-width: 35px;
+}
+
+.audio-elements-container {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    gap: 15px;
+    min-height: 100px;
+    padding: 20px;
+    border: 2px dashed #bdc3c7;
+    border-radius: 8px;
+    background-color: #f8f9fa;
+}
+
+.audio-element {
+    padding: 12px 16px;
+    background: #e3f2fd;
+    border: 2px solid #2196f3;
+    border-radius: 6px;
+    cursor: grab;
+    transition: all 0.3s ease;
+    font-weight: 500;
+    color: #1976d2;
+    min-width: 120px;
+    text-align: center;
+    position: relative;
+    user-select: none;
+}
+
+.audio-element:hover {
+    background: #bbdefb;
+    transform: translateY(-2px);
+    box-shadow: 0 4px 12px rgba(33, 150, 243, 0.3);
+}
+
+.audio-element.dragging {
+    opacity: 0.7;
+    cursor: grabbing;
+    transform: rotate(5deg);
+}
+
+.audio-element.playing {
+    background: #fff3e0;
+    border-color: #ff9800;
+    color: #e65100;
+}
+
+.audio-element .play-icon {
+    display: inline-block;
+    margin-right: 8px;
+    font-size: 16px;
+    cursor: pointer;
+    padding: 6px;
+    border-radius: 50%;
+    background: rgba(255, 255, 255, 0.3);
+    transition: all 0.2s ease;
+    width: 28px;
+    height: 28px;
+    text-align: center;
+    line-height: 16px;
+}
+
+.audio-element .play-icon:hover {
+    background: rgba(255, 255, 255, 0.5);
+    transform: scale(1.1);
+}
+
+.groups-container {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+    gap: 20px;
+    margin-top: 30px;
+}
+
+.group {
+    background: white;
+    border-radius: 10px;
+    padding: 20px;
+    box-shadow: 0 2px 10px rgba(0,0,0,0.1);
+    min-height: 200px;
+}
+
+.group-header {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-bottom: 15px;
+    padding-bottom: 10px;
+    border-bottom: 2px solid #ecf0f1;
+}
+
+.group-image {
+    width: 120px;
+    height: 120px;
+    border-radius: 12px;
+    object-fit: cover;
+    border: 3px solid #e9ecef;
+    box-shadow: 0 4px 8px rgba(0,0,0,0.1);
+}
+
+.drop-zone {
+    min-height: 120px;
+    border: 2px dashed #bdc3c7;
+    border-radius: 8px;
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    padding: 15px;
+    transition: all 0.3s ease;
+}
+
+.drop-zone.drag-over {
+    border-color: #3498db;
+    background-color: #ebf3fd;
+}
+
+.audio-slot {
+    padding: 12px 16px;
+    background: #f8f9fa;
+    border: 2px dashed #bdc3c7;
+    border-radius: 6px;
+    text-align: center;
+    color: #6c757d;
+    transition: all 0.3s ease;
+    min-width: 120px;
+    min-height: 50px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+}
+
+.audio-slot.drag-over {
+    border-color: #3498db;
+    background-color: #ebf3fd;
+    border-style: solid;
+}
+
+.audio-slot.filled {
+    background: #6c757d;
+    border: 2px solid #495057;
+    color: white;
+    border-style: solid;
+    cursor: grab;
+    position: relative;
+}
+
+.audio-slot.filled:hover {
+    background: #5a6268;
+    transform: translateY(-1px);
+}
+
+.audio-slot.filled.dragging {
+    opacity: 0.7;
+    cursor: grabbing;
+    transform: rotate(3deg);
+}
+
+.audio-slot.filled .play-button {
+    position: absolute;
+    top: 2px;
+    right: 2px;
+    width: 20px;
+    height: 20px;
+    background: rgba(255, 255, 255, 0.2);
+    border: none;
+    border-radius: 50%;
+    color: white;
+    cursor: pointer;
+    font-size: 10px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    transition: all 0.2s ease;
+}
+
+.audio-slot.filled .play-button:hover {
+    background: rgba(255, 255, 255, 0.4);
+    transform: scale(1.1);
+}
+
+.audio-slot.filled.correct {
+    background: #28a745;
+    border-color: #1e7e34;
+}
+
+.audio-slot.filled.incorrect {
+    background: #dc3545;
+    border-color: #bd2130;
+}
+
+.audio-slot.disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+}
+
+.controls {
+    text-align: center;
+    margin: 30px 0;
+}
+
+.btn {
+    padding: 12px 24px;
+    border: none;
+    border-radius: 6px;
+    font-size: 16px;
+    cursor: pointer;
+    margin: 0 10px;
+    transition: all 0.3s ease;
+}
+
+.btn-primary {
+    background: #3498db;
+    color: white;
+}
+
+.btn-primary:hover {
+    background: #2980b9;
+}
+
+.btn-success {
+    background: #27ae60;
+    color: white;
+}
+
+.btn-success:hover {
+    background: #229954;
+}
+
+@media (max-width: 768px) {
+    .groups-container {
+        grid-template-columns: 1fr;
+    }
+    
+    .title {
+        font-size: 2rem;
+    }
+}
